LengthLayer: dtype casting

+++ TFNetworkLayer.py
@@ -1974,16 +1974,16 @@ class LengthLayer(LayerBase):
   """
   layer_class = "length"
 
-  def __init__(self, add_time_axis=False, **kwargs):
+  def __init__(self, add_time_axis=False, dtype="int32", **kwargs):
     super(LengthLayer, self).__init__(**kwargs)
     assert len(self.sources) == 1, "%s: expects one source" % self
-    out = tf.cast(self.sources[0].output.size_placeholder[self.sources[0].output.time_dim_axis], tf.int32)
+    out = tf.cast(self.sources[0].output.size_placeholder[self.sources[0].output.time_dim_axis], dtype)
     if add_time_axis:
       out = tf.expand_dims(out, axis=self.output.time_dim_axis)
     self.output.placeholder = out
 
   @classmethod
-  def get_out_data_from_opts(cls, name, sources, add_time_axis=False, **kwargs):
+  def get_out_data_from_opts(cls, name, sources, add_time_axis=False, dtype="int32", **kwargs):
     if add_time_axis:
       shape = (1,)
       time_dim_axis = 1
@@ -1995,7 +1995,7 @@ class LengthLayer(LayerBase):
       shape=shape,
       batch_dim_axis=0,
       time_dim_axis=time_dim_axis,
-      dtype="int32",
+      dtype=dtype,
       sparse=False)
